refactor: move route options func and test into route/options package
